+++ portlets/lecture2go-portlet/docroot/WEB-INF/liferay-display.xml
@@ -9,7 +9,6 @@
 		<portlet id="lg-admin-terms-management"/>
 		<portlet id="lg-admin-lecture-series-management"/>
         <portlet id="lg-admin-institution-management"></portlet>
-        <portlet id="lg-admin-scheduler-management"></portlet>
         <portlet id="lg-admin-thread-management"></portlet>
         <portlet id="lg-open-access-videos"></portlet>
         <portlet id="lg-admin-category-management"></portlet>
