+++ portlets/lecture2go-portlet/docroot/WEB-INF/liferay-portlet.xml
@@ -115,7 +115,7 @@
 	</portlet>
 	
 	<portlet>
-   	 <portlet-name>lg-admin-scheduler-management</portlet-name>
+   	 <portlet-name>lg-admin-thread-management</portlet-name>
    	 	<icon>/icon.png</icon>
    	 	<scheduler-entry> 
    	 	    <scheduler-description>
