+++ portlets/lecture2go-portlet/docroot/WEB-INF/portlet.xml
@@ -206,12 +206,12 @@
   </portlet>
   
   <portlet>
-        <portlet-name>lg-admin-scheduler-management</portlet-name>
-        <display-name>L2Go Scheduler Management</display-name>
-        <portlet-class>de.uhh.l2g.plugins.admin.SchedulerManagement</portlet-class>
+        <portlet-name>lg-admin-thread-management</portlet-name>
+        <display-name>L2Go Threads Management</display-name>
+        <portlet-class>de.uhh.l2g.plugins.admin.ThreadsManagement</portlet-class>
         <init-param>
             <name>view-template</name>
-            <value>/admin/jobs.jsp</value>
+            <value>/admin/threads.jsp</value>
         </init-param>
         <init-param>
             <name>add-process-action-success-action</name>
@@ -224,9 +224,9 @@
         </supports>
         <resource-bundle>content.Language</resource-bundle>
         <portlet-info>
-            <title>L2Go Scheduler Management</title>
-            <short-title>L2Go Scheduler Management</short-title>
-            <keywords>L2Go Scheduler Management</keywords>
+            <title>L2Go Threads Management</title>
+            <short-title>L2Go Threads Management</short-title>
+            <keywords>L2Go Threads Management</keywords>
         </portlet-info>
   </portlet>
