+++ portlets/lecture2go-portlet/docroot/WEB-INF/src/de/uhh/l2g/plugins/admin/ThreadManagement.java
@@ -26,4 +26,232 @@ public class ThreadManagement extends MVCPortlet {
 
 	
 
+
+	/**Set default permissions (assumes fixed and unique role names)
+	 * 
+	 * @param pm - PermissionManager 
+	 * @throws PortalException 
+	 * @throws SystemException 
+	 */
+	private void setDefaultPermissions(PermissionManager pm) throws SystemException, PortalException{
+		
+
+	}
+	
+	@Override
+	public void render(RenderRequest renderRequest, RenderResponse renderResponse) throws PortletException, IOException {
+        
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+			         ScheduledThread.class.getName(), renderRequest);
+        
+	
+	    } catch (Exception e) {
+	    	throw new PortletException(e);
+	    }
+
+		super.render(renderRequest, renderResponse);
+	 
+	}
+	
+	public void startJob(ActionRequest request, ActionResponse response){
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			 
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+			 
+			  //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  scheduler.start();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	
+	public void resumeJob(ActionRequest request, ActionResponse response){
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			 
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+			 
+			  //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  scheduler.resume();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	public void pauseJob(ActionRequest request, ActionResponse response){		
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+				
+			 //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  
+			  scheduler.pause();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+			
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	
+	public void updateJob(ActionRequest request, ActionResponse response){		
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+				
+			 //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  
+			  scheduler.update();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+			
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	public void stopJob(ActionRequest request, ActionResponse response){		
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+				
+			 //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  
+			  scheduler.stop();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+			
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	
+	public void removeJob(ActionRequest request, ActionResponse response){		
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+				
+			 //Make sure to use the appropriate Message Consumer 
+			  ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  
+			  scheduler.initScheduler(schedulerClassName, serviceContext.getPortletId());
+			  System.out.println(scheduler.getPortletId());
+			  
+			  scheduler.stop();
+
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+			
+			System.out.println(e.getClass().getName());
+			e.printStackTrace();
+			
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	
+	public void removeAllJobs(ActionRequest request, ActionResponse response){
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+					ScheduledThread.class.getName(), request);
+			
+			  String schedulerClassName = ParamUtil.getString(request, "schedulerName");
+			  if (schedulerClassName.isEmpty()){
+				  PortletScheduler.removeAllPortletSchedulerJobs();
+			  }
+			  else{	//Use the correct Message Consumer 
+			  	ClassLoader classLoader = PortletClassLoaderUtil.getClassLoader(serviceContext.getPortletId()); //Where portletID is not null
+			  	PortletScheduler scheduler = (PortletScheduler) classLoader.loadClass(schedulerClassName).newInstance();
+			  
+			  	scheduler.removeAllJobs();
+			  }
+		 } catch (Exception e) {
+			SessionErrors.add(request, e.getClass().getName());
+			PortalUtil.copyRequestParameters(request, response);
+
+			response.setRenderParameter("mvcPath",
+					"/admin/jobs.jsp");
+		}
+	}
+	
+	/**
+	 * 
+	 * TODO: Remove duplicates from List before the can be scheduled, or correctly remove allready unscheduled Entries
+	 */
+	public void init() throws PortletException{	
+		super.init();
+		
+		/*Drop table LG_VideoStatistics and initialize view
+		 * 
+		 */
+	}
+	
+	public void destroy(){
+	//	System.out.println(this.getPortletName());
+	//	PortletScheduler.ListSchedulerEntries(this.getPortletName());
+		super.destroy();
+	}
+	
 }
