+++ portlets/lecture2go-portlet/docroot/WEB-INF/src/service.properties
@@ -13,8 +13,8 @@
 ##
 
     build.namespace=LG
-    build.number=1390
-    build.date=1461168907129
+    build.number=1394
+    build.date=1461262795547
     build.auto.upgrade=true
 
 ##
